move S2Graph global helpers into PostProcess.
